remove mergeState which is no longer used
